requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
Flask==2.2.0
Flask-Session==0.8.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
numpy==2.0.2
pandas>=2.0.0
pyodbc==5.1.0
python-dateutil>=2.8.2
python-dotenv==1.0.1
pytz==2024.2
requests>=2.28.1
six==1.17.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==2.2.0
gunicorn==21.2.0
Flask-CORS>=4.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
APScheduler>=3.10.4